refactor(codey-mac): extract shared flow graph renderer with status styling
Moves WorkerNodeView, ConditionNodeView, TerminalNodeView, FlowEdgeView,
nodeTypes, edgeTypes, EDGE_COLOR, resolveColor, and rfNodeType out of
FlowEditor.tsx into a new flowGraph.tsx shared module. Adds optional
data.status styling (border/glow/icon) to WorkerNodeView for the forthcoming
TeamRunFlow read-only view — when status is undefined (FlowEditor's case)
rendering is byte-for-byte identical to before.
