+++ codey-mac/src/components/FlowEditor.tsx
@@ -1,127 +1,14 @@
 import { useCallback, useMemo, useState } from 'react'
 import {
   ReactFlow, Background, Controls, addEdge, applyNodeChanges, applyEdgeChanges,
-  ReactFlowProvider, useReactFlow,
-  Handle, Position, NodeResizer, ConnectionMode,
-  BaseEdge, EdgeLabelRenderer, getBezierPath, MarkerType,
-  type Node, type Edge, type Connection, type NodeProps, type EdgeProps,
+  ReactFlowProvider, useReactFlow, ConnectionMode, MarkerType,
+  type Node, type Edge, type Connection,
 } from '@xyflow/react'
 import '@xyflow/react/dist/style.css'
 import { TeamGraph, validateGraph } from '../../../packages/core/src/team-graph'
 import { toFlow, fromFlow, newNodeId } from './flowEditorModel'
+import { nodeTypes, edgeTypes, EDGE_COLOR, resolveColor, rfNodeType } from './flowGraph'
 import { C, useEffectiveTheme } from '../theme'
-
-// A single edge color throughout — edges don't encode meaning by color; the
-// branch (yes/no) and conditions are shown as labels instead.
-const EDGE_COLOR = C.accent
-
-// SVG <marker> fill is a presentation attribute — CSS var() references don't
-// resolve there. Read the actual computed hex from the document root instead.
-const resolveColor = (cssVar: string) =>
-  getComputedStyle(document.documentElement).getPropertyValue(cssVar.slice(4, -1)).trim()
-
-// ---------------------------------------------------------------------------
-// Custom node components
-// ---------------------------------------------------------------------------
-
-const HANDLE_IDS = ['t', 'r', 'b', 'l'] as const
-const HANDLE_POS = { t: Position.Top, r: Position.Right, b: Position.Bottom, l: Position.Left }
-
-// One handle per side. With ConnectionMode.Loose (set on <ReactFlow>), a single
-// `source` handle can both start and receive a connection, so a drag can begin
-// from any side. Rendering separate source+target handles per side stacked them
-// on top of each other — the target covered the source, so drags could never
-// start and no edges could be drawn.
-function NodeHandles() {
-  return (
-    <>
-      {HANDLE_IDS.map(id => (
-        <Handle key={id} type="source" id={id} position={HANDLE_POS[id]} style={{ width: 9, height: 9, background: C.accent }} />
-      ))}
-    </>
-  )
-}
-
-// A node flagged by validation gets a red border; the selected node gets an
-// accent ring. `bad` is injected per-render from the validation problems.
-function ring(selected?: boolean, bad?: boolean): string | undefined {
-  if (selected) return `0 0 0 3px ${C.accent}, 0 0 14px 2px ${C.accent}`
-  if (bad) return `0 0 0 1px ${C.red}`
-  return undefined
-}
-
-function WorkerNodeView({ data, selected }: NodeProps) {
-  const d = data as { label: string; role?: string; bad?: boolean }
-  // Size to content (full name on one line); never clip the name. React Flow
-  // applies any user-resized width/height to the node wrapper itself, so we
-  // don't bind them here — that previously collapsed the card and clipped text.
-  return (
-    <div style={{ minWidth: 100, padding: '8px 12px', borderRadius: 8, background: C.surface2, border: `1px solid ${d.bad ? C.red : C.border}`, color: C.fg, boxSizing: 'border-box', boxShadow: ring(selected, d.bad) }}>
-      <NodeResizer isVisible={selected} minWidth={100} minHeight={44} />
-      <div style={{ fontSize: 13, fontWeight: 600, whiteSpace: 'nowrap' }}>{d.label}</div>
-      {d.role && <div style={{ fontSize: 11, color: C.fg2, marginTop: 2, maxWidth: 240, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>{d.role}</div>}
-      <NodeHandles />
-    </div>
-  )
-}
-
-function ConditionNodeView({ data, selected }: NodeProps) {
-  const d = data as { condition?: string; bad?: boolean }
-  const text = d.condition?.trim() || 'condition?'
-  return (
-    <div style={{ width: 130, height: 130, display: 'flex', alignItems: 'center', justifyContent: 'center', position: 'relative' }}>
-      {/* The diamond is a rotated square behind upright, centered text. */}
-      <div style={{ position: 'absolute', inset: 19, transform: 'rotate(45deg)', background: C.surface2, border: `1px solid ${d.bad ? C.red : C.accent}`, borderRadius: 8, boxShadow: ring(selected, d.bad) }} />
-      <span style={{ position: 'relative', fontSize: 10, lineHeight: 1.25, color: C.fg, width: 78, textAlign: 'center', display: '-webkit-box', WebkitLineClamp: 5, WebkitBoxOrient: 'vertical', overflow: 'hidden' }}>{text}</span>
-      <NodeHandles />
-    </div>
-  )
-}
-
-function TerminalNodeView({ data, selected }: NodeProps) {
-  const d = data as { label: string; bad?: boolean }
-  return (
-    <div style={{ padding: '10px 22px', borderRadius: 999, background: C.bg, border: `1px solid ${d.bad ? C.red : C.border}`, color: C.fg, fontSize: 14, fontWeight: 600, boxShadow: ring(selected, d.bad) }}>
-      {d.label}
-      <NodeHandles />
-    </div>
-  )
-}
-
-const nodeTypes = { workerNode: WorkerNodeView, conditionNode: ConditionNodeView, terminalNode: TerminalNodeView }
-
-// Custom edge: single color, optional label, and — when `data.running` is set
-// (the edge is reachable from the start node) — a dot that travels along the
-// path to visualize flow leaving start.
-function FlowEdgeView({ id, sourceX, sourceY, targetX, targetY, sourcePosition, targetPosition, markerEnd, label, data, selected }: EdgeProps) {
-  const [edgePath, labelX, labelY] = getBezierPath({ sourceX, sourceY, targetX, targetY, sourcePosition, targetPosition })
-  const running = (data as any)?.running
-  // Edges in the live flow (reachable from start) use the accent color; dangling
-  // edges are greyed. The selected edge is always accent and drawn thicker.
-  const stroke = selected ? C.accent : running ? EDGE_COLOR : C.fg2
-  return (
-    <>
-      {selected && <path d={edgePath} fill="none" stroke={C.accent} strokeWidth={9} strokeLinecap="round" style={{ opacity: 0.25 }} />}
-      <BaseEdge id={id} path={edgePath} markerEnd={markerEnd} style={{ stroke, strokeWidth: selected ? 3.5 : 1.5 }} />
-      {running && (
-        <circle r={4} fill={stroke}>
-          <animateMotion dur="1.6s" repeatCount="indefinite" path={edgePath} rotate="auto" />
-        </circle>
-      )}
-      {label != null && label !== '' && (
-        <EdgeLabelRenderer>
-          <div style={{
-            position: 'absolute', transform: `translate(-50%, -50%) translate(${labelX}px, ${labelY}px)`,
-            background: C.bg, color: C.fg2, fontSize: 10, padding: '1px 5px', borderRadius: 4,
-            border: `1px solid ${C.border}`, pointerEvents: 'none',
-          }}>{String(label)}</div>
-        </EdgeLabelRenderer>
-      )}
-    </>
-  )
-}
-
-const edgeTypes = { flowEdge: FlowEdgeView }
 
 // ---------------------------------------------------------------------------
 // Props
@@ -140,7 +27,7 @@ function FlowEditorInner({ teamName, workerNames, workerRoles = {}, graph, onSav
   const effectiveTheme = useEffectiveTheme()
   const withTypes = (ns: Node[]): Node[] => ns.map(n => {
     const t = (n.data as any).type
-    const rfType = t === 'worker' ? 'workerNode' : t === 'condition' ? 'conditionNode' : 'terminalNode'
+    const rfType = rfNodeType(t)
     const role = t === 'worker' ? workerRoles[(n.data as any).worker] : undefined
     return { ...n, type: rfType, data: { ...n.data, role } }
   })
